# Optimize the providers tests when running subsets of them