--- a/tests/providers/apache/drill/operators/test_drill.py
+++ b/tests/providers/apache/drill/operators/test_drill.py
@@ -17,6 +17,7 @@
 # under the License.
 
 import unittest
+from unittest import mock
 
 import pytest
 
@@ -52,7 +53,7 @@ class TestDrillOperator(unittest.TestCase):
         select * from cp.`employee.json` limit 10
         """
         op = DrillOperator(task_id='drill_operator_test_single', sql=sql, dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
+        op.execute(mock.MagicMock())
 
     def test_drill_operator_multi(self):
         sql = [
@@ -60,4 +61,4 @@ class TestDrillOperator(unittest.TestCase):
             "select sum(employee_id), any_value(full_name)" "from dfs.tmp.test_airflow",
         ]
         op = DrillOperator(task_id='drill_operator_test_multi', sql=sql, dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
+        op.execute(mock.MagicMock())