--- a/tests/providers/postgres/operators/test_postgres.py
+++ b/tests/providers/postgres/operators/test_postgres.py
@@ -17,10 +17,10 @@
 # under the License.
 
 import unittest
+from unittest import mock
 
 import pytest
 
-from airflow.models.dag import DAG
 from airflow.providers.postgres.operators.postgres import PostgresOperator
 from airflow.utils import timezone
 
@@ -32,11 +32,6 @@ TEST_DAG_ID = 'unit_test_dag'
 
 @pytest.mark.backend("postgres")
 class TestPostgres(unittest.TestCase):
-    def setUp(self):
-        args = {'owner': 'airflow', 'start_date': DEFAULT_DATE}
-        dag = DAG(TEST_DAG_ID, default_args=args)
-        self.dag = dag
-
     def tearDown(self):
         tables_to_drop = ['test_postgres_to_postgres', 'test_airflow']
         from airflow.providers.postgres.hooks.postgres import PostgresHook
@@ -52,13 +47,11 @@ class TestPostgres(unittest.TestCase):
             dummy VARCHAR(50)
         );
         """
-        op = PostgresOperator(task_id='basic_postgres', sql=sql, dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
+        op = PostgresOperator(task_id='basic_postgres', sql=sql)
+        op.execute(mock.MagicMock())
 
-        autocommit_task = PostgresOperator(
-            task_id='basic_postgres_with_autocommit', sql=sql, dag=self.dag, autocommit=True
-        )
-        autocommit_task.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
+        autocommit_task = PostgresOperator(task_id='basic_postgres_with_autocommit', sql=sql, autocommit=True)
+        autocommit_task.execute(mock.MagicMock())
 
     def test_postgres_operator_test_multi(self):
         sql = [
@@ -66,8 +59,8 @@ class TestPostgres(unittest.TestCase):
             "TRUNCATE TABLE test_airflow",
             "INSERT INTO test_airflow VALUES ('X')",
         ]
-        op = PostgresOperator(task_id='postgres_operator_test_multi', sql=sql, dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
+        op = PostgresOperator(task_id='postgres_operator_test_multi', sql=sql)
+        op.execute(mock.MagicMock())
 
     def test_vacuum(self):
         """
@@ -75,8 +68,8 @@ class TestPostgres(unittest.TestCase):
         """
 
         sql = "VACUUM ANALYZE;"
-        op = PostgresOperator(task_id='postgres_operator_test_vacuum', sql=sql, dag=self.dag, autocommit=True)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
+        op = PostgresOperator(task_id='postgres_operator_test_vacuum', sql=sql, autocommit=True)
+        op.execute(mock.MagicMock())
 
     def test_overwrite_schema(self):
         """
@@ -94,7 +87,5 @@ class TestPostgres(unittest.TestCase):
 
         from psycopg2 import OperationalError
 
-        try:
-            op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-        except OperationalError as e:
-            assert 'database "foobar" does not exist' in str(e)
+        with pytest.raises(OperationalError, match=r'.+database "foobar" does not exist.+'):
+            op.execute(mock.MagicMock)