Some tests take an excessively long time to run when you run a subset of the tests locally. This is because they use BaseOperator.run () which loads a lot of classes. In many cases, we can call the BaseOperator.execute method directly, which speeds up these tests.
For example:
Before:
```
$  pytest tests/providers/sqlite/ | tail -n 1
================================== 11 passed, 6 warnings in 9.18s ===================================
$ pytest tests/providers/neo4j/ | tail -n 1
======================== 5 passed, 6 warnings in 9.55s =========================
```

After:
```
$  pytest tests/providers/sqlite/ | tail -n 1
======================================== 11 passed in 1.99s =========================================
pytest tests/providers/neo4j/ | tail -n 1
============================== 5 passed in 1.55s ===============================
```
<!--
Thank you for contributing! Please make sure that your code changes
are covered with tests. And in case of new features or big changes
remember to adjust the documentation.

Feel free to ping committers for the review!

In case of existing issue, reference it using one of the following:

closes: #ISSUE
related: #ISSUE

How to write a good git commit message:
http://chris.beams.io/posts/git-commit/
-->

---
**^ Add meaningful description above**

Read the **[Pull Request Guidelines](https://github.com/apache/airflow/blob/main/CONTRIBUTING.rst#pull-request-guidelines)** for more information.
In case of fundamental code change, Airflow Improvement Proposal ([AIP](https://cwiki.apache.org/confluence/display/AIRFLOW/Airflow+Improvements+Proposals)) is needed.
In case of a new dependency, check compliance with the [ASF 3rd Party License Policy](https://www.apache.org/legal/resolved.html#category-x).
In case of backwards incompatible changes please leave a note in [UPDATING.md](https://github.com/apache/airflow/blob/main/UPDATING.md).